requirements.txt
aiosmtpd==1.4.6
asgiref==3.9.1
atpublic==6.0.1
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.6
defusedxml==0.7.1
Django==5.2.4
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
easy-thumbnails==2.10
idna==3.10
MarkupSafe==3.0.2
oauthlib==3.3.1
pillow==11.3.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python3-openid==3.2.0
redis==6.4.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
